Summary
The original number of rows and columns in the data was [330 rows x 27 columns]. The remaining rows and columns in the data are [277 rows x 25 columns].
The Big Mountain Resort was present in the data with no missing values.
The fastEight column was dropped in its entirety because half the values were missing and all but the others are the value zero. Thus there was essentially no information in this column.
The Adult Weekday prices columns were dropped since between Weekend and Weekday prices, Adult Weekend prices were preferred because there were less missing values.
The row for Heavenly snowmaking area covered was dropped because the value was much higher than expected and no ticket pricing information was available at all for this resort so any further effort about values for this resort will be wasted.
14% of the rows contained no price data in either Adult Weekend or Adult Weekday pricing and thus were dropped since the target is Price data.
One row for resort years open was dropped because the value was 2019 years. Unsure whether this resort in question has been open zero years or one year and even whether the numbers are projections or actual. In any case, you would be adding a new youngest resort so it feels best to simply drop this row.

Data correction: The skiiable terrain area for Silverton Mountain resort was suspiciously high at a value of 26819. Data was corrected to 1819 was the value found after looking up. The last three digits agree. Error suspectd to have occured in transmission or some editing or transcription stage. You could plausibly replace the suspect value with the one you've just obtained.

State-wide summary statistics for our market segment were added to the data to see if state-wide supply and demand of certain skiing resources, available night skiing, total skiiable terrain, or days open may factor into pricing strategies. US Population and area data obtained from Wikipedia was added into the summary statistics to help explore the relationship between US states.

The target feature to predict ticket pricing is weekend prices in Montana resorts.  Observations show that there is a clear line where weekend and weekday prices are equal. Overall, weekend prices being higher than weekday prices seem restricted to sub $100 resorts. However, the distribution for weekday and weekend prices in Montana seemed equal.